Convert ChartSettingsTableFormatting components to TS (#54103)
