+++ frontend/src/metabase/visualizations/components/settings/ChartSettingsTableFormatting/ChartSettingsTableFormatting.unit.spec.tsx
@@ -2,9 +2,14 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { useState } from "react";
 
+import { render, screen } from "__support__/ui";
+import type { ColumnFormattingSetting } from "metabase-types/api";
 import { createMockColumn } from "metabase-types/api/mocks";
 
-import { ChartSettingsTableFormatting } from "./ChartSettingsTableFormatting";
+import {
+  ChartSettingsTableFormatting,
+  type ChartSettingsTableFormattingProps,
+} from "./ChartSettingsTableFormatting";
 
 const STRING_COLUMN = createMockColumn({
   base_type: "type/Text",
@@ -77,8 +82,8 @@ const NUMBER_OPERATORS = [
 
 const BOOLEAN_OPERATORS = ["is null", "is not null", "is true", "is false"];
 
-const Wrapper = props => {
-  const [value, setValue] = useState([]);
+const Wrapper = (props: Partial<ChartSettingsTableFormattingProps> = {}) => {
+  const [value, setValue] = useState<ColumnFormattingSetting[]>([]);
 
   return (
     <ChartSettingsTableFormatting
@@ -90,7 +95,7 @@ const Wrapper = props => {
   );
 };
 
-const setup = props => {
+const setup = (props = {}) => {
   render(<Wrapper {...props} />);
 };
