+++ frontend/src/metabase/visualizations/components/settings/ChartSettingsTableFormatting/index.ts
@@ -1,1 +1,3 @@
-export * from "./ChartSettingsTableFormatting";
+export { ChartSettingsTableFormatting } from "./ChartSettingsTableFormatting";
+export { isFormattable } from "./util";
+export { ALL_OPERATOR_NAMES } from "./get-operators-for-columns";
