+++ frontend/src/metabase/visualizations/lib/table_format.unit.spec.js
@@ -1,4 +1,4 @@
-import { ALL_OPERATOR_NAMES } from "metabase/visualizations/components/settings/ChartSettingsTableFormatting/get-operators-for-columns";
+import { ALL_OPERATOR_NAMES } from "metabase/visualizations/components/settings/ChartSettingsTableFormatting";
 import {
   OPERATOR_FORMATTER_FACTORIES,
   canCompareSubstrings,
